Add XML documentation to manager classes
Added detailed XML summary, remarks, and parameter documentation to AudioManager, BoardManager, HealthManager, ItemManager, ScoreManager, and SnakeManager classes. This improves code maintainability and developer understanding by clarifying the purpose and usage of methods, properties, signals, and data structures throughout the domain manager scripts.
