+++ Scripts/Domain/Managers/ItemManager.cs
@@ -25,30 +25,161 @@ using Snake.Scripts.Interfaces;
 
 namespace Snake.Scripts.Domain.Managers;
 
+/// <summary>
+/// Manages game items, including their spawn rates, placement, interactions, and effects within the game board.
+/// Signals out impact of different items to Main.
+/// </summary>
 public partial class ItemManager : Node
 {
-	[Signal] public delegate void HealthDeductedRequestedEventHandler();
-	[Signal] public delegate void ScoreChangedEventHandler(double amount, bool isDelta);
-	[Signal] public delegate void ComboPointsXChangedEventHandler(double amount, bool isDelta);
-	[Signal] public delegate void ComboPointsYChangedEventHandler(double amount, bool isDelta);
-	[Signal] public delegate void ComboStartedEventHandler();
-	[Signal] public delegate void ComboEndedEventHandler();
-	[Signal] public delegate void ComboCancelledEventHandler();
-	[Signal] public delegate void GameOverRequestedEventHandler();
-	[Signal] public delegate void HudFlashRequestedEventHandler(int type);
-	[Signal] public delegate void ItemSpawnedEventHandler(Node2D item);
-	[Signal] public delegate void EggMovedEventHandler();
-	[Signal] public delegate void BadFoodEatenEventHandler();
-	[Signal] public delegate void SpecialEggEatenEventHandler(bool isInCombo);
-	[Signal] public delegate void HighChimeRequestedEventHandler();
+	/// <summary>
+	/// Represents the delegate for handling events triggered when a health deduction is requested.
+	/// This signal communicates the details of the health deduction process.
+	/// </summary>
+	[Signal]
+	public delegate void HealthDeductedRequestedEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when the score changes.
+	/// This signal provides the amount of change and indicates whether the change is incremental.
+	/// </summary>
+	[Signal]
+	public delegate void ScoreChangedEventHandler(double amount, bool isDelta);
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when the X-coordinate of combo points changes.
+	/// This signal provides details about the amount of change and indicates if the change is incremental.
+	/// </summary>
+	[Signal]
+	public delegate void ComboPointsXChangedEventHandler(double amount, bool isDelta);
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when the Y-coordinate of combo points changes.
+	/// This signal provides details about the amount of change and specifies if the change is incremental.
+	/// </summary>
+	[Signal]
+	public delegate void ComboPointsYChangedEventHandler(double amount, bool isDelta);
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when a combo sequence has started.
+	/// This signal informs the system to initiate or respond to the beginning of a combo action.
+	/// </summary>
+	[Signal]
+	public delegate void ComboStartedEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when a combo sequence has successfully ended.
+	/// This signal notifies the system to respond to the completion of an ongoing combo action.
+	/// </summary>
+	[Signal]
+	public delegate void ComboEndedEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when a combo sequence is interrupted or canceled.
+	/// This signal notifies the system to respond to the cancellation of an ongoing combo action.
+	/// </summary>
+	[Signal]
+	public delegate void ComboCancelledEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events that request the termination of the game session.
+	/// This signal is used to notify the system when certain conditions are met that warrant ending the game.
+	/// </summary>
+	[Signal]
+	public delegate void GameOverRequestedEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events that request a flash effect on the HUD.
+	/// This signal is used to instruct the HUD to visually indicate certain actions or events
+	/// based on a specified type.
+	/// </summary>
+	[Signal]
+	public delegate void HudFlashRequestedEventHandler(int type);
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when an item is spawned.
+	/// This signal is used to notify subscribers about the creation or appearance
+	/// of a new item within the game environment.
+	/// </summary>
+	[Signal]
+	public delegate void ItemSpawnedEventHandler(Node2D item);
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when an egg is moved.
+	/// This signal is used to notify subscribers about changes in the position
+	/// of an egg within the game environment.
+	/// </summary>
+	[Signal]
+	public delegate void EggMovedEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when bad food is eaten.
+	/// This signal is used to notify subscribers about the consumption of an undesirable food item
+	/// and can be used to implement negative effects or penalties in the game.
+	/// </summary>
+	[Signal]
+	public delegate void BadFoodEatenEventHandler();
+
+	/// <summary>
+	/// Represents the delegate for handling events triggered when a special egg is eaten.
+	/// This signal is used to notify subscribers about the special egg consumption,
+	/// optionally including whether it is part of a combo sequence.
+	/// </summary>
+	[Signal]
+	public delegate void SpecialEggEatenEventHandler(bool isInCombo);
+
+	/// <summary>
+	/// Represents the delegate for handling events when a high chime is requested.
+	/// This signal is intended to notify listeners to execute actions associated with the request.
+	/// </summary>
+	[Signal]
+	public delegate void HighChimeRequestedEventHandler();
 	
 	private readonly ISnakeManager _snake;
+
+	/// <summary>
+	/// Represents the size of a cell in pixels.
+	/// This property is used to define the pixel dimensions of an individual cell
+	/// within the context of the item's management system.
+	/// </summary>
 	public int CellPixelSizeRef { get; set; }
+
+	/// <summary>
+	/// Defines the reference size of a board 'cell', meaning a square within the grid.
+	/// This property is used to establish the dimensions of a cell within the board's coordinate system.
+	/// </summary>
 	public int BoardCellSizeRef { get; set; }
+
+	/// <summary>
+	/// Defines the offset to be applied to the placement of items on the game board.
+	/// This property is used to adjust the position of game elements to ensure
+	/// consistent and visually aligned placement relative to the grid and other elements.
+	/// </summary>
 	public Vector2I PlacementOffsetRef { get; set; }
+
+	/// <summary>
+	/// Represents the node associated with the egg in the game.
+	/// This property is used to manage the egg's visual and positional state on the board.
+	/// </summary>
 	public Node2D EggNode { get; set; }
+
+	/// <summary>
+	/// Represents a reference to the current score.
+	/// This property is used to track and manage the score within the system.
+	/// </summary>
 	public double CurrentScoreRef { get; set; }
+
+	/// <summary>
+	/// Tracks the current combo multiplier reference value.
+	/// This property is used to maintain and update the multiplier points
+	/// for combinations in the context of the item's management system.
+	/// </summary>
 	public double CurrentComboPointsXRef { get; set; }
+
+	/// <summary>
+	/// Represents the Y-position reference for the current combo points.
+	/// This property is used to determine or track the vertical position
+	/// of the combo points within the system.
+	/// </summary>
 	public double CurrentComboPointsYRef { get; set; }
 
 	internal Node2D WallNode;
@@ -92,17 +223,35 @@ public partial class ItemManager : Node
 	internal Node2D LargeWallNode;
 	private ItemEffectLogic _effectLogic;
 	private double _eggPulseTime;
+	private double _itemPulseTime;
 	private Vector2I _offset;
 
+	/// <summary>
+	/// Manages the operations and logic related to items within the game, including their placement,
+	/// effects, and interactions. This class handles item spawning, item effects, and interactions
+	/// such as scoring, combo mechanics, wall placements, and egg movements.
+	/// It includes event signaling for game updates and interactions.
+	/// </summary>
+	/// <param name="snake">The snake manager instance responsible for snake-related operations and interactions.</param>
 	public ItemManager(ISnakeManager snake)
 	{
 		_snake = snake;
 	}
 
+	/// <summary>
+	/// Parameterless initialization of a new instance of the ItemManager class.
+	/// </summary>
 	public ItemManager()
 	{
 	}
 
+	/// <summary>
+	/// Loads and initializes item nodes and their corresponding file paths for the game.
+	/// This method retrieves item-related nodes from the provided item manager and maps them
+	/// to internal node fields. It also extracts file paths for each item's corresponding scene
+	/// and prepares essential logic for item effects.
+	/// </summary>
+	/// <param name="itemManager">The node containing item-related child nodes and configurations.</param>
 	internal void LoadItems(Node itemManager)
 	{
 		WallNode = itemManager.GetNode<Node2D>("Wall");
@@ -137,7 +286,14 @@ public partial class ItemManager : Node
 
 		_effectLogic = new ItemEffectLogic(_snake, _rottenEggPath, _dewDropPath);
 	}
-	
+
+	/// <summary>
+	/// Configures item spawn rates by mapping item categories to their respective node representations.
+	/// This method initializes a dictionary where each key represents a spawn frequency, and its corresponding value
+	/// is a list of items to be spawned at that frequency. Higher keys denote rarer items.
+	/// The mapping ensures that different nodes, such as eggs, walls, and power-ups, are assigned varying likelihoods
+	/// of appearing in the game. This configuration influences item distribution dynamics during gameplay.
+	/// </summary>
 	internal void SetItemRates()
 	{
 		_itemRates = new Dictionary<int, List<Node2D>>
@@ -159,6 +315,13 @@ public partial class ItemManager : Node
 	}
 
 
+	/// <summary>
+	/// Randomly repositions the egg to a new valid spot on the game board, ensuring it is not trapped by walls.
+	/// This method identifies all available positions on the board and selects one at random.
+	/// It verifies that the chosen position does not result in a wall trap. If a trap is detected,
+	/// it continues to select new positions until a valid one is found. Once the egg's new position
+	/// is determined, its visual representation is updated accordingly.
+	/// </summary>
 	internal void MoveEgg()
 	{
 		EmitSignal(SignalName.EggMoved);
@@ -177,25 +340,43 @@ public partial class ItemManager : Node
 		}
 	}
 
+	/// <summary>
+	/// Checks if the current egg position is completely surrounded by walls, forming a trap.
+	/// Specifically, this method evaluates whether the North, East, South, and West neighboring cells
+	/// relative to the egg's position are all occupied by walls. If all these neighboring cells contain walls,
+	/// the position is considered trapped and unsuitable for the egg's placement.
+	/// </summary>
+	/// <returns>
+	/// Returns <c>true</c> if the egg's position is fully enclosed by walls in all four cardinal directions,
+	/// indicating a wall trap. Otherwise, returns <c>false</c>.
+	/// </returns>
 	private bool CheckWallTrap()
 	{
-		//TODO:Doesn't account for large walls, or larger perimeter traps... it should, but I wonder if it's necessary or worth it. A certain planned power-up might make it even less so
 		var eggPosNord = new Vector2I(EggPosition.X, EggPosition.Y - 1);
 		var eggPosEas = new Vector2I(EggPosition.X + 1, EggPosition.Y);
 		var eggPosSud = new Vector2I(EggPosition.X, EggPosition.Y + 1);
 		var eggPosWes = new Vector2I(EggPosition.X - 1, EggPosition.Y);
+		// ReSharper disable once ConvertIfStatementToReturnStatement
 		if (WallsData.Contains(eggPosNord) && WallsData.Contains(eggPosEas) && WallsData.Contains(eggPosSud) &&
-			WallsData.Contains(eggPosWes))
+		    WallsData.Contains(eggPosWes))
 		{
 			return true;
 		}
 		return false;
 	}
 
+	/// <summary>
+	/// Randomly selects a valid position for item placement on the board.
+	/// The method ensures that the selected position complies with several constraints, such as avoiding
+	/// collisions with existing items, large items, or the snake's body, and maintaining a specified minimum
+	/// distance from the snake's head and a certain proximity to the snake's tail.
+	/// </summary>
+	/// <param name="available">A list of available positions on the board where items can potentially be placed.</param>
+	/// <returns>Returns a <c>Vector2I</c> object representing the valid randomly-selected position for placement.</returns>
 	private Vector2I RandomPlacement(List<Vector2I> available)
 	{
 		Random rndm = new Random();
-		
+
 		int tryCounter = 0;
 		
 		Vector2I itemPlacement;
@@ -208,7 +389,6 @@ public partial class ItemManager : Node
 				EmitSignal(SignalName.GameOverRequested);
 			}
 			itemPlacement = new Vector2I(rndm.Next(1, BoardCellSizeRef + 1), rndm.Next(3, BoardCellSizeRef - 4)); 
-			//TODO: more dynamic way to factor for GUI frame in offsets?
 			
 		} while ((!available.Contains(itemPlacement) || //Don't place on an occupied position.
 				  CheckLargeItemHit(itemPlacement) || //Don't place on large 
@@ -218,8 +398,7 @@ public partial class ItemManager : Node
 				 ));
 
 		return itemPlacement;
-
-		//TODO: Needs a better implementation. (*Why? Don't remember what is wrong with this.*)
+		
 		bool CheckWithinRadius(Vector2I newItemPlacement, List<Vector2I> listCoords, int radius)
 		{
 			return listCoords.Any(spot => IsWithinRadius(newItemPlacement, spot, radius));
@@ -234,6 +413,14 @@ public partial class ItemManager : Node
 		}
 	}
 
+	/// <summary>
+	/// Checks if the given position intersects with any part of a large item on the board.
+	/// A large item is represented as a cluster of cells forming a square.
+	/// The method verifies if the position matches the top-left corner of the cluster or
+	/// any of its adjacent cells (right, bottom, or bottom-right).
+	/// </summary>
+	/// <param name="position">The position on the board to check for a collision with a large item.</param>
+	/// <returns>Returns true if the position overlaps any part of a large item; otherwise, false.</returns>
 	internal bool CheckLargeItemHit(Vector2I position)
 	{
 		bool hit = false;
@@ -252,6 +439,12 @@ public partial class ItemManager : Node
 		return hit;
 	}
 
+	/// <summary>
+	/// Generates new items based on the current tally and predefined item rate rules.
+	/// For each rate rule, items are spawned if the tally satisfies the rule's condition.
+	/// New items are created by duplicating existing item templates and are placed
+	/// on the game board in valid positions.
+	/// </summary>
 	internal void GenerateFromItemLookup()
 	{
 		// ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
@@ -265,6 +458,11 @@ public partial class ItemManager : Node
 		}
 	}
 
+	/// <summary>
+	/// Places the specified item on the game board in a valid position based
+	/// on its type, ensuring it does not overlap with existing items, walls, or other occupied areas.
+	/// </summary>
+	/// <param name="newItem">The item to be placed on the game board, represented as a Node2D object.</param>
 	private void SpawnItem(Node2D newItem)
 	{
 		newItem.Visible = true;
@@ -310,6 +508,15 @@ public partial class ItemManager : Node
 		EmitSignal(SignalName.ItemSpawned, newItem);
 	}
 
+	/// <summary>
+	/// Identifies and returns all available positions on the game board where items can be placed,
+	/// ensuring that the specified area does not overlap with existing snake segments, items, walls,
+	/// or other occupied positions.
+	/// </summary>
+	/// <param name="area">The size of the area to check for availability, represented as a Vector2I
+	/// where X and Y define the width and height of the area respectively.</param>
+	/// <returns>A list of Vector2I positions representing the top-left corners of all available
+	/// spots on the game board that can accommodate the specified area.</returns>
 	private List<Vector2I> FindAvailableSpots(Vector2I area)
 	{
 		List<Vector2I> available = [];
@@ -355,6 +562,13 @@ public partial class ItemManager : Node
 		return available;
 	}
 
+	/// <summary>
+	/// Processes the result of an item interaction, handling its removal, triggering signals,
+	/// and applying its effects based on the item's type and whether it is part of a combo.
+	/// </summary>
+	/// <param name="item">The item being processed, represented as a Node2D instance.</param>
+	/// <param name="i">The index of the item in the corresponding item data list.</param>
+	/// <param name="isInCombo">A boolean indicating if the item interaction is part of a combo sequence.</param>
 	internal void ItemResult(Node2D item, int i, bool isInCombo)
 	{
 		if (item.SceneFilePath != LargeWallNode.SceneFilePath)
@@ -382,6 +596,13 @@ public partial class ItemManager : Node
 		ApplyItemEffect(item.SceneFilePath, isInCombo);
 	}
 
+	/// <summary>
+	/// Applies the effect of the specified item based on its type and whether it is part of a combo.
+	/// Depending on the item's properties, this may modify the snake's attributes, update the score,
+	/// emit signals, or trigger special game events.
+	/// </summary>
+	/// <param name="sceneFilePath">The file path representing the item's scene, used to identify the type of item.</param>
+	/// <param name="isInCombo">A boolean indicating if the item is part of a combo sequence, which may alter the effect logic.</param>
 	internal void ApplyItemEffect(string sceneFilePath, bool isInCombo)
 	{
 		_effectLogic?.Apply(sceneFilePath);
@@ -597,6 +818,10 @@ public partial class ItemManager : Node
 		}
 	}
 
+	/// <summary>
+	/// Handles the logic executed when the egg is eaten by the snake. This includes updating the tally of eaten eggs,
+	/// repositioning the egg on the game board, and generating a new set of items at randomized locations.
+	/// </summary>
 	public void EggEaten()
 	{
 		Tally++;
@@ -604,6 +829,13 @@ public partial class ItemManager : Node
 		GenerateFromItemLookup();
 	}
 
+	/// <summary>
+	/// Updates the pulse animation and rotation for the egg. The pulse effect modifies the egg's
+	/// scale based on a sinusoidal pattern, creating a growing and shrinking effect. Additionally,
+	/// when the combo mode is active, the egg rotates to enhance the animation.
+	/// </summary>
+	/// <param name="delta">The time elapsed since the last frame, in seconds.</param>
+	/// <param name="isCombo">Indicates whether the combo mode is currently active.</param>
 	public void UpdateEggPulse(double delta, bool isCombo)
 	{
 		if (EggNode == null) return;
@@ -622,8 +854,16 @@ public partial class ItemManager : Node
 		}
 	}
 
+	/// <summary>
+	/// Updates the pulse animation for all items. This method adjusts the scaling effect
+	/// based on a sinusoidal animation pattern. The pulse effect is only applied when
+	/// the combo mode is active.
+	/// </summary>
+	/// <param name="delta">The elapsed time interval since the last frame, in seconds.</param>
+	/// <param name="isCombo">A boolean value indicating whether the combo mode is currently active.</param>
 	public void UpdateItemPulse(double delta, bool isCombo)
 	{
+		_itemPulseTime += delta;
 		const float cycleTime = (float)3.0;
 		if (!isCombo) return;
 		
@@ -631,11 +871,17 @@ public partial class ItemManager : Node
 		foreach (var item in ItemNodes)
 		{
 			variance += 0.1;
-			float pulseFactor = (float)(1.0 + 0.25 * Math.Sin(2.0 * Math.PI * _eggPulseTime / (cycleTime + variance)));
+			float pulseFactor = (float)(1.0 + 0.25 * Math.Sin(2.0 * Math.PI * _itemPulseTime / (cycleTime + variance)));
 			item.Scale = new Vector2(pulseFactor, pulseFactor);
 		}
 	}
 
+	/// <summary>
+	/// Releases all resources used by the item manager, including clearing all
+	/// item and wall-related data, freeing all associated nodes, and resetting
+	/// the necessary collections to their initial states. Additionally, updates
+	/// the egg's position after resource disposal.
+	/// </summary>
 	public new void Dispose()
 	{
 		Tally = 0;
@@ -683,6 +929,13 @@ public partial class ItemManager : Node
 		MoveEgg();
 	}
 
+	/// <summary>
+	/// Updates the score and combo point references used to track the player's progress
+	/// in the game. These values are updated based on the current state of the gameplay.
+	/// </summary>
+	/// <param name="score">The updated score to be assigned to the CurrentScoreRef property.</param>
+	/// <param name="comboX">The updated X component of combo points to be assigned to the CurrentComboPointsXRef property.</param>
+	/// <param name="comboY">The updated Y component of combo points to be assigned to the CurrentComboPointsYRef property.</param>
 	public void UpdateScoreReferences(double score, double comboX, double comboY)
 	{
 		CurrentScoreRef = score;
@@ -690,23 +943,43 @@ public partial class ItemManager : Node
 		CurrentComboPointsYRef = comboY;
 	}
 
+	/// <summary>
+	/// Sets the cell pixel size reference, defining the size of each cell
+	/// in pixels for accurate positioning on the game board.
+	/// </summary>
+	/// <param name="boardCellPixelSize">The size of a single cell in pixels, specified as an integer.</param>
 	public void SetCellPixelSizeRef(int boardCellPixelSize)
 	{
 		CellPixelSizeRef = boardCellPixelSize;
 	}
 
+	/// <summary>
+	/// Sets the board cell size reference, which determines the size of each cell
+	/// on the game board in logically defined units.
+	/// </summary>
+	/// <param name="boardBoardCellSize">The size of a single logical cell on the board, specified as an integer.</param>
 	public void SetBoardCellSizeRef(int boardBoardCellSize)
 	{
 		BoardCellSizeRef = boardBoardCellSize;
 	}
 
+	/// <summary>
+	/// Sets the placement offset for item placement operations,
+	/// which determines the reference point for positioning items on the game board.
+	/// </summary>
+	/// <param name="vector2I">The offset in pixels specified as a Vector2I value.</param>
 	public void SetPlacementOffset(Vector2I vector2I)
 	{
 		PlacementOffsetRef = vector2I;
 	}
 
+	/// <summary>
+	/// Initializes the EggNode with the specified Node2D instance,
+	/// allowing it to be used for egg-related operations within the item management system.
+	/// </summary>
+	/// <param name="getNode">The Node2D instance to assign as the EggNode.</param>
 	public void InitializeEggNode(Node2D getNode)
 	{
-		EggNode	= getNode;
+		EggNode = getNode;
 	}
 }
