+++ Scripts/Domain/Managers/TimeManager.cs
@@ -24,12 +24,21 @@ using Timer = Godot.Timer;
 
 namespace Snake.Scripts.Domain.Managers;
 
+/// <summary>
+/// Manages game logic timers such as movement updates, HUD flashing, and health updates in the game.
+/// </summary>
 public partial class TimeManager : GodotObject
 {
     private Timer _moveTimer;
     private Timer _hudFlashTimer;
     private Timer _healthTimer;
 
+    /// <summary>
+    /// Initializes the TimeManager by setting the provided timer instances for movement, HUD flashing, and health updates.
+    /// </summary>
+    /// <param name="moveTimer">The timer responsible for controlling movement updates.</param>
+    /// <param name="hudFlashTimer">The timer responsible for managing periodic HUD flashing actions.</param>
+    /// <param name="healthTimer">The timer responsible for managing periodic health-related updates.</param>
     public void Initialize(Timer moveTimer, Timer hudFlashTimer, Timer healthTimer)
     {
         _moveTimer = moveTimer;
@@ -37,14 +46,52 @@ public partial class TimeManager : GodotObject
         _healthTimer = healthTimer;
     }
 
+    /// <summary>
+    /// Starts the move timer, enabling periodic movement updates in the game logic.
+    /// </summary>
     public void StartMoveTimer() => _moveTimer.Start();
+
+    /// <summary>
+    /// Stops the move timer, halting the periodic movement updates in the game logic.
+    /// </summary>
     public void StopMoveTimer() => _moveTimer.Stop();
+
+    /// <summary>
+    /// Sets the wait time interval for the move timer, controlling the delay between movements in the game logic.
+    /// </summary>
+    /// <param name="waitTime">The desired wait time interval for the move timer, in seconds.</param>
     public void SetMoveTimerWaitTime(double waitTime) => _moveTimer.WaitTime = waitTime;
+
+    /// <summary>
+    /// Retrieves the current wait time interval configured for the move timer.
+    /// Use this to determine the delay between movements in the game logic.
+    /// </summary>
+    /// <returns>The wait time interval of the move timer in seconds.</returns>
     public double GetMoveTimerWaitTime() => _moveTimer.WaitTime;
 
+    /// <summary>
+    /// Initiates the HUD flash timer, activating time-based flashing functionality
+    /// associated with it. Use this to enable recurring HUD flash effects
+    /// dependent on the timer's operation.
+    /// </summary>
     public void StartHudFlashTimer() => _hudFlashTimer.Start();
+
+    /// <summary>
+    /// Stops the HUD flash timer, disabling any active time-based flashing functionality
+    /// associated with it. Use this to halt the recurring HUD flash effects that depend
+    /// on the timer's operation.
+    /// </summary>
     public void StopHudFlashTimer() => _hudFlashTimer.Stop();
 
+    /// <summary>
+    /// Starts the health timer, enabling time-based logic or updates associated with health.
+    /// Use this to initiate recurring health-related events driven by the timer.
+    /// </summary>
     public void StartHealthTimer() => _healthTimer.Start();
+
+    /// <summary>
+    /// Stops the health timer, preventing further timeout signals from being emitted.
+    /// This can be used to halt health-related updates or logic tied to the timer.
+    /// </summary>
     public void StopHealthTimer() => _healthTimer.Stop();
 }
